+++ tests/clickhousegrant_test.go
@@ -364,7 +364,7 @@ func TestClickhouseGrantExample(t *testing.T) {
 	s := NewSession(ctx, k8sClient, cfg.Project)
 
 	// Cleans test afterward
-	defer s.Destroy()
+	defer s.Destroy(t)
 
 	// WHEN
 	// Applies given manifest