+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Ignore `http.StatusBadRequest` on `ClickhouseGrant` deletion
 - Retry conflict error when k8s object saved to the storage
+- Fix `ClickhouseGrant` invalid remote and local privileges comparison
 
 ## v0.22.0 - 2024-07-02