+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Ignore `http.StatusBadRequest` on `ClickhouseGrant` deletion
 - Retry conflict error when k8s object saved to the storage
+- Fix `ClickhouseGrant` invalid remote and local privileges comparison
+- Fix `ClickhouseGrant`: doesn't escape role name to grant
 
 ## v0.22.0 - 2024-07-02